WebApp/API/requirements.txt
fastapi==0.75.1
happytransformer==2.4.1
nltk==3.6.5
pandas==1.3.4
pydantic==1.8.2
pyocclient==0.6
sql_metadata==2.4.0
starlette==0.17.1
tqdm==4.62.3
uvicorn==0.17.6
nltk
word2number
scispacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl